perplexity-mcp-server/requirements.txt
# Core
asyncio
aiohttp
httpx>=0.24.0
pydantic>=2.0
python-dotenv

# JSON-RPC
jsonrpc-base==2.1.0
jsonrpc-async==2.1.0

# Testing
pytest
pytest-asyncio
pytest-mock